requirements.txt
streamlit
supabase
google-generative-ai
Pillow